+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 <!-- ### Enhancements -->
 <!-- ### Backwards incompatible changes -->
 
+## Unreleased
+
+- Add `Nimrag.steps_weekly`
+- Add `Nimrag.sleep_daily`
+- Add `Nimrag.user_settings`
+
 ## 0.1.0 (2024-03-29)
 
 Init release